+++ packages/holz-stream-backend/src/index.ts
@@ -1,0 +1,1 @@
+export { default, default as StreamBackend } from './stream-backend';